More code-review comments fixed
